* 'master' of github.com:tlmquintino/configs:
  better install
  made path general

+++ install.sh
@@ -3,17 +3,20 @@
 cd $HOME
 
 # bash setup
-ln -sf .myconfigs/bash/bash_profile .bash_profile
-ln -sf .myconfigs/bash/bashrc .bashrc
+[[ ! -f .bash_profile ]] && ln -sf .myconfigs/bash/bash_profile .bash_profile
+[[ ! -f .bashrc ]] && ln -sf .myconfigs/bash/bashrc .bashrc
 
 # git setup
-ln -sf .myconfigs/git/gitconfig .gitconfig
-ln -sf .myconfigs/git/gitignore .gitignore
+[[ ! -f .gitconfig ]] && ln -sf .myconfigs/git/gitconfig .gitconfig
+[[ ! -f .gitignore ]] && ln -sf .myconfigs/git/gitignore .gitignore
 
 # vim setup
-ln -sf .myconfigs/vim .vim
-ln -sf .myconfigs/vim/vimrc .vimrc
-ln -sf .myconfigs/vim/gvimrc .gvimrc
+[[ ! -f .vim ]] && ln -sf .myconfigs/vim .vim
+[[ ! -f .vimrc ]] && ln -sf .myconfigs/vim/vimrc .vimrc
+[[ ! -f .gvimrc ]] && ln -sf .myconfigs/vim/gvimrc .gvimrc
 
 # emacs setup
-ln -sf .myconfigs/emacs/emacs .emacs
+[[ ! -f .emacs ]] && ln -sf .myconfigs/emacs/emacs .emacs
+
+# create local profile
+[[ ! -f .myconfigs/extra/$HOSTNAME/bash/profile ]] && mkdir -p .myconfigs/extra/$HOSTNAME/bash && touch .myconfigs/extra/$HOSTNAME/bash/profile
